Removido a mensagem de sucesso que aparecia quando um user novo entrava no site pela primeira vez. Modificado mensagem devolvida ao user caso ele edite seu perfil. Adicionado link para retornar ao MEU PERFIL caso o user edite seu perfil. Se a descricao for nula, imprime uma mensagem informando o user disso. Adicionado uma linha em branco no perfil. Ao pedir um convite para um grupo, a mensagem de erro nao eh mais exibida. Um user que pediu convite, ao olhar para o perfil do time que pediu o convite, vera uma mensagem informando que jah tem um convite enviado. Um user comum pode agora deixar o time. A mensagem de O USER FOI EXPULSO ao deletar a participacao do grupo foi removida. Feito gambiarra na classe userTeam para poder excluir um user sem dar erro. Adicionado forma de ver o perfil do team depois de sair do grupo.

+++ app/classes/UserTeam.php
@@ -164,15 +164,20 @@ class UserTeam extends Model {
 	 *************************************************************************/
 	public function delete($idUser, $idTeam) 
 	{
+	
+		if($idUser == "" || $idTeam == "")
+		{
+			return False;
+		}
+	/*
 		if ($this->checkId($this->table_name, $idUser) && $this->checkId($this->table_name, $idTeam)) 
 		{
 			$this->messageFail("Não existe tal user registrado nesse grupo.");
 			return False;
-		}
+	*/	
 		else
 		{
 			$this->db->query("DELETE FROM " . $this->table_name . " WHERE idUser = $idUser AND idTeam = $idTeam");
-			$this->messageOk("O usuário foi expulso com sucesso!");
 			return True;
 		}
 	}
